Move static styles to stylesheet
I also took the liberty of changing the colors of the bar chart and the
tooltip.

There are parts that I couldn't move, like the chart's width and
height, but those values are used in subsequent calculations, so I just
left them there (though you can get those values programmatically, but
I got weird results (maybe because those values are returned as
strings, but I'll just leave it at that)).