Split the style imports into their own bundle
So I added another entry point just for the style imports, because it
doesn't make sense to import styles on the main application logic. I'm
not sure about the load order of multiple bundles, but it looks like
the styles bundle is loaded after the main bundle (though it looks like
the load order doesn't matter here).